Function page_allocator::allocate_pages_by_bytes_deferred

pub fn allocate_pages_by_bytes_deferred(
    requested_vaddr: Option<VirtualAddress>,
    num_bytes: usize
) -> Result<(AllocatedPages, DeferredAllocAction<'static>), &'static str>

Similar to allocated_pages_deferred(), but accepts a size value for the allocated pages in number of bytes instead of number of pages.

This function still allocates whole pages by rounding up the number of bytes.