Function page_allocator::allocate_pages_deferred

pub fn allocate_pages_deferred(
    requested_vaddr: Option<VirtualAddress>,
    num_pages: usize,
    within_range: Option<&PageRange>
) -> Result<(AllocatedPages, DeferredAllocAction<'static>), &'static str>

The core page allocation routine that allocates the given number of virtual pages, optionally at the requested starting VirtualAddress.

This simply reserves a range of virtual addresses, it does not allocate actual physical memory frames nor do any memory mapping. Thus, the returned AllocatedPages aren’t directly usable until they are mapped to physical frames.

Allocation is based on a red-black tree and is thus O(log(n)). Fragmentation isn’t cleaned up until we’re out of address space, but that’s not really a big deal.

Arguments

requested_vaddr: if Some, the returned AllocatedPages will start at the Page containing this VirtualAddress. If None, the first available Page range will be used, starting at any random virtual address.

num_pages: the number of Pages to be allocated.

Return

If successful, returns a tuple of two items:

the pages that were allocated, and

an opaque struct representing details of bookkeeping-related actions that may cause heap allocation. Those actions are deferred until this returned DeferredAllocAction struct object is dropped, allowing the caller (such as the heap implementation itself) to control when heap allocation may occur.